scripts/requirements.txt
python-frontmatter>=1.0
